Scrolling Displays and Model Fixes

+++ fabric/src/main/java/org/mtr/mod/client/DynamicTextureCache.java
@@ -75,7 +75,7 @@ public class DynamicTextureCache implements IGui {
 		deletedResourcesToRemove.forEach(deletedResources::removeLong);
 	}
 
-	public DynamicResource getPixelatedText(String text, int textColor, int maxWidth, float cjkSizeRatio, boolean fullPixel) {
+	public DynamicResource getPixelatedText(String text, int textColor, int maxWidth, double cjkSizeRatio, boolean fullPixel) {
 		return getResource(String.format("pixelated_text_%s_%s_%s_%s_%s", text, textColor, maxWidth, cjkSizeRatio, fullPixel), () -> RouteMapGenerator.generatePixelatedText(text, textColor, maxWidth, cjkSizeRatio, fullPixel), DefaultRenderingColor.TRANSPARENT);
 	}
